requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2
python-dotenv==1.0.0
cmdstanpy==1.1.3
prophet==1.1.5
scikit-learn==1.3.2